Rename the scripts folder to be more pythonic; update the imports to the new package.
